:sparkles: :: 애그리거트 root 클래스에 애그리거트 합치기
