+++ crowdfunding/src/main/kotlin/com/project/indistraw/crowdfunding/adapter/input/CrowdfundingWebAdapter.kt
@@ -6,18 +6,11 @@ import com.project.indistraw.crowdfunding.adapter.input.data.response.Crowdfundi
 import com.project.indistraw.crowdfunding.adapter.input.data.response.CrowdfundingListResponse
 import com.project.indistraw.crowdfunding.adapter.input.data.response.CrowdfundingPagingResponse
 import com.project.indistraw.crowdfunding.application.port.input.*
-import com.project.indistraw.crowdfunding.domain.StatusType
 import org.springframework.data.domain.Pageable
 import org.springframework.data.web.PageableDefault
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PatchMapping
-import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RestController
+import org.springframework.web.bind.annotation.*
 
 @RestController
 @RequestMapping("/api/v1/crowdfunding")
@@ -28,7 +21,6 @@ class CrowdfundingWebAdapter(
     private val crowdfundingListUseCase: CrowdfundingListUseCase,
     private val popularCrowdfundingListUseCase: PopularCrowdfundingListUseCase,
     private val findMyCrowdfundingListUseCase: FindMyCrowdfundingListUseCase,
-    private val updateCrowdfundingStatusUseCase: UpdateCrowdfundingStatusUseCase
 ) {
 
     @PostMapping
@@ -59,10 +51,5 @@ class CrowdfundingWebAdapter(
         findMyCrowdfundingListUseCase.execute()
             .let { crowdfundingDataConverter.toResponse(it) }
             .let { ResponseEntity.ok(it) }
-
-    @PatchMapping("/{idx}/status/{statusType}")
-    fun updateCrowdfundingStatus(@PathVariable idx: Long, @PathVariable statusType: StatusType): ResponseEntity<Void> =
-        updateCrowdfundingStatusUseCase.execute(idx, statusType)
-            .let { ResponseEntity.status(HttpStatus.RESET_CONTENT).build() }
 
 }
