+++ crowdfunding/src/main/kotlin/com/project/indistraw/crowdfunding/adapter/input/data/response/CrowdfundingDetailResponse.kt
@@ -1,6 +1,6 @@
 package com.project.indistraw.crowdfunding.adapter.input.data.response
 
-import com.project.indistraw.crowdfunding.domain.StatusType
+import com.project.indistraw.crowdfunding.domain.Crowdfunding
 import java.util.*
 
 data class CrowdfundingDetailResponse(
@@ -14,7 +14,7 @@ data class CrowdfundingDetailResponse(
     val imageList: List<String>,
     val fileList: List<String>,
     val reward: List<RewardResponse>,
-    val status: StatusType
+    val status: Crowdfunding.StatusType
 ) {
 
     data class Writer(
