+++ crowdfunding/src/main/kotlin/com/project/indistraw/crowdfunding/adapter/output/persistence/repository/CrowdfundingRepository.kt
@@ -1,7 +1,7 @@
 package com.project.indistraw.crowdfunding.adapter.output.persistence.repository
 
 import com.project.indistraw.crowdfunding.adapter.output.persistence.entity.CrowdfundingEntity
-import com.project.indistraw.crowdfunding.domain.StatusType
+import com.project.indistraw.crowdfunding.domain.Crowdfunding
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
 import org.springframework.data.jpa.repository.JpaRepository
@@ -9,8 +9,8 @@ import java.util.*
 
 interface CrowdfundingRepository: JpaRepository<CrowdfundingEntity, Long> {
 
-    fun findTop5ByStatusTypeNotOrderByViewCountDesc(statusType: StatusType): List<CrowdfundingEntity>
+    fun findTop5ByStatusTypeNotOrderByViewCountDesc(statusType: Crowdfunding.StatusType): List<CrowdfundingEntity>
     fun findByWriterIdx(writerIdx: UUID): List<CrowdfundingEntity>
-    fun findAllByStatusTypeNot(statusType: StatusType, pageable: Pageable): Page<CrowdfundingEntity>
+    fun findAllByStatusTypeNot(statusType: Crowdfunding.StatusType, pageable: Pageable): Page<CrowdfundingEntity>
 
 }
