--- crowdfunding/src/main/kotlin/com/project/indistraw/crowdfunding/domain/Amount.kt
@@ -1,9 +1,0 @@
-package com.project.indistraw.crowdfunding.domain
-
-import com.project.indistraw.crowdfunding.application.common.annotation.Aggregate
-
-@Aggregate
-data class Amount(
-    val totalAmount: Long,
-    val targetAmount: Long
-)
