+++ crowdfunding/src/main/kotlin/com/project/indistraw/crowdfunding/domain/CrowdfundingViewCount.kt
@@ -1,5 +1,8 @@
 package com.project.indistraw.crowdfunding.domain
 
+import com.project.indistraw.crowdfunding.application.common.annotation.AggregateRoot
+
+@AggregateRoot
 data class CrowdfundingViewCount(
     val crowdfundingIdx: Long,
     val ips: Set<String>
