--- crowdfunding/src/main/kotlin/com/project/indistraw/crowdfunding/domain/DirectorAccount.kt
@@ -1,9 +1,0 @@
-package com.project.indistraw.crowdfunding.domain
-
-import com.project.indistraw.crowdfunding.application.common.annotation.Aggregate
-
-@Aggregate
-data class DirectorAccount(
-    val bank: String,
-    val account: String
-)
