--- crowdfunding/src/main/kotlin/com/project/indistraw/crowdfunding/domain/StatusType.kt
@@ -1,10 +1,0 @@
-package com.project.indistraw.crowdfunding.domain
-
-import com.project.indistraw.crowdfunding.application.common.annotation.Aggregate
-
-@Aggregate
-enum class StatusType {
-
-    UNDER_REVIEW, RECRUITING, FINISHED
-
-}
